fix: break App.tsx↔screen circular dependency causing StarSwarm TDZ crash
Move RootStackParamList / HomeStackParamList / ProfileStackParamList into
frontend/src/types/navigation.ts. All 21 screens now import from that
leaf module instead of from App.tsx, eliminating the cycle:

  App.tsx → lazyScreens.ts → StarSwarmScreen.tsx → App.tsx

App.tsx re-exports the three types for any code still referencing them
from their old location.

Fixes #1553
