+++ frontend/src/screens/MahjongLayoutInspectorScreen.tsx
@@ -3,7 +3,7 @@ import { FlatList, Pressable, StyleSheet, Text, View } from "react-native";
 import { useNavigation } from "@react-navigation/native";
 import { NativeStackNavigationProp } from "@react-navigation/native-stack";
 import { useSafeAreaInsets } from "react-native-safe-area-context";
-import type { HomeStackParamList } from "../../App";
+import type { HomeStackParamList } from "../types/navigation";
 import { LAYOUTS } from "../game/mahjong/layouts/registry";
 import type { LayoutMeta } from "../game/mahjong/types";
 import { GameShell } from "../components/shared/GameShell";
