+++ frontend/src/screens/ProfileScreen.tsx
@@ -16,7 +16,7 @@ import { useTheme } from "../theme/ThemeContext";
 import { AppHeader, APP_HEADER_HEIGHT } from "../components/shared/AppHeader";
 import { statsApi } from "../api/stats";
 import type { StatsResponse, GameRow } from "../api/types";
-import type { ProfileStackParamList } from "../../App";
+import type { ProfileStackParamList } from "../types/navigation";
 import { formatDate } from "../utils/formatTimestamp";
 
 type ProfileNav = NativeStackNavigationProp<ProfileStackParamList, "ProfileHome">;
